Update index.md
This is my suggestion of a beginning of this wiki page. Accoring to the wiki's ToDo, there should be more than just a copy of the most recent default-config.
I've tried to explain those passages that I as a (new) user found most intriguing to change or most demanding to understand when first using fvwm and trying to tweak it a bit to my personal needs and ideas.
Let's see what you all think before I send a pull request to the wiki's original master branch.
Cheers!