SCSS: rename files to avoid conflict with jekyll
It seems something has changed in jekyll or its dependencies, such that,
without this chane `jekyll serve` fails:

jekyll/converters/scss.rb:175:in `rescue in convert': expected "{".
(Jekyll::Converters::Scss::SyntaxError)

It is unclear precisely why this change is required, but it looks as
though it disambiguates between the same-named files used elsewhere.