Cookbook/AltTab: SwitchWindow to use infostore
Now that fvwm has had infostore for sometime, change the example of the
SwitchWindow function to make use of it.  This avoids the need for
PipeRead, and hence using the shell purely for environment expansion.

Add a new SwitchDirection to reduce complexity, remove the need for
a second variable, reduce the number of calls to the conditionals.

Noticed by @andibachmann on Github.  SwitchWindow changes mostly from
@somiaj.