pages: add Dockerfile
WIP -- this a quick PoC for creating a small environment which can build
a jekyll website, but serve it from within the container.

Github pages is deployed in the container so we can tie that publishing
step into this, once the site is being built how we want.

To create the container first of all, it needs to be built:

```
docker build -t fvwm-website .
```

Once that's done, you can run it from the top-level git repo:

```
docker run -v $(pwd)/:/home/site -w /home/site -p 4000:4000 fvwm-website jekyll serve --host 0.0.0.0
```

This will bind mount the files from PWD into the container, as
/home/site -- and then run jekyll serve in the container.

Go to:  http://localhost:4000

To see the rendered results.