Add syntax highlighting plugin using Pygments.
This adds my old Pygments lexer before Jekyll moved to rouge. This is done by using
a Jekyll plugin to call the `pygmentize` binary to convert the code using the lexer
`_FvwmLexer.py` included in the source. Because the plugin is run for each block of
code, a full python environment is loaded each time `pygmentize` is called, which
increases the build times significantly (30seconds on my machine). By default the
plugin will not run `pygmentize`, so this feature needs to be enabled by running:

jekyll build --config _config.yml,_config-fvwm2rc.yml

Since the site is built rarely, implementing a faster method is not work any additional
complexity in using the custom lexer. Also since this is a custom _plugin, this will
not work on githubpages directly.