Add fvwm3 manual pages.
This adds the fvwm3 manual pages back to the website due to demand.
The majority of the output is created directly by asciidoctor with a
shell script that I've submitted to the fvwm3 source that slightly modifies
it to better fit on the website. This is the current manual pages as
of the 1.1.0 release.