CI: add custom build container
This moves the jekyll build and gh-pages deployment stage into a
customer Github actions file.

Doing this will allow for us to use any custom jekyll plugins which
can't be used with Github's native build system for GH pages, and
therefore we only need to upload the result of the site generation.