Add script _build_fvwm3_man.sh to build fvwm3 manpages.
This script takes a single input, the location of the fvwm3 source,
and then builds all the manual pages and puts them in Man/. Run script
from top level of this source.

    ./_build_fvwm3_man.sh /path/to/fvwm3_source

This script is meant to be used with the fvwm3 release CI scripts
to update the manual pages at that time.